// src/client/App/constants/Constants.ts
export enum ActionTypes {
    FETCH_MOVIES_SUCCESS = "FETCH_MOVIES_SUCCESS",
    FETCH_MOVIES_PENDING = "FETCH_MOVIES_PENDING",
    FETCH_MOVIES_ERROR = "FETCH_MOVIES_ERROR",
    SET_SEARCH_INPUT = "SET_SEARCH_INPUT",
    SET_SEARCH_FILTER = "SET_SEARCH_FILTER",
}

export enum LocalStorageKeys {
    LAST_SEARCH = "LAST_SEARCH"
}